Release 1.1
+ Added ability to read and subscribe to Lighthouse SDK server
settings, in particular whether the SDK should be enabled or not. When
the SDK is disabled non of the SDK commands will perform functionality.
This means you can include the SDK in a release of your app with it
disabled on the server and then in the future you can update the server
to enabled and the SDK will begin to perform desired functionality.

+ Add parameter while request push notifications. Developer can use its own send id to implement GCM push notification.